% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplyLulu.R
\name{ApplyLulu}
\alias{ApplyLulu}
\title{ApplyLulu}
\usage{
ApplyLulu(seqs, table, output, vsearchdest = "vsearch")
}
\arguments{
\item{seqs}{The location of a FASTA file containing sequences used in the generation of the OTU-sample table.}

\item{table}{The location of a .csv sample table to be used for lulu OTU pruning. Must contain the same OTU name as the supplied FASTA seqs.}

\item{output}{An output file location and name for the lulu pruned OTU-sample table .csv file.}

\item{vsearchdest}{Specify the location of the vsearch executable if PATH not configured appropriately.}
}
\value{
None
}
\description{
This function automates the lulu algorithm on a user supplied set of sequences and sample-OTU mapping.
}
