% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dadaReadPrep.R
\name{dadaReadPrep}
\alias{dadaReadPrep}
\title{dadaReadPrep}
\usage{
dadaReadPrep(PrimerF = NA, PrimerR = NA, UsePrimerFile = FALSE,
  folderwfiles = "1.rawreads", folderoutput = "7.DADA2/trimmed.reads",
  lookforsecondprimer = FALSE, cutadaptdest = "cutadapt", ncores = 1)
}
\arguments{
\item{PrimerF}{Sequence of the forward primer}

\item{PrimerR}{Sequence of the reverse primer}

\item{UsePrimerFile}{TRUE or FALSE, is a csv file provided containing primer information?}

\item{folderwfiles}{A directory indicating files to be processed if not default folder.}

\item{folderoutput}{Output directory.}

\item{lookforsecondprimer}{TRUE or FALSE, should Cutadapt look for a linked primer in the sequence, this is useful if the single read length covers the entire amplicon.}

\item{cutadaptdest}{Full directory for the cutadapt executable if not in PATH.}

\item{ncores}{Number of processsor cores to use.}
}
\value{
None
}
\description{
This function prepares paired-end Illumina data for the DADA2 pipeline.
It takes in paired-end demultiplexed amplicon fastq data and primer data and uses Cutadapt to trim off the primer region.
Input files are expected to have the structure SampleID_S*_L001_R1_001.fastq.
}
