% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastqCount.R
\name{FastqCount}
\alias{FastqCount}
\title{FastqCount}
\usage{
FastqCount(file)
}
\arguments{
\item{file}{The file location.}
}
\value{
The number of reads in the fastq file
}
\description{
This function counts the number of reads in a fastq file. It handles one file at a time.
}
\examples{
FastqCount(sample)
sapply(manyfiles,FastqCount)

}
