% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OTUCluster.R
\name{OTUCluster}
\alias{OTUCluster}
\title{OTUCluster}
\usage{
OTUCluster(UsePrimerFile = FALSE, usearchdest = "usearch")
}
\arguments{
\item{UsePrimerFile}{TRUE or FALSE, is a csv file provided containing primer information?}

\item{usearchdest}{Specify the location of the usearch executable if PATH not configured appropriately.}
}
\value{
None
}
\description{
This clusters sequences at 97% similarity to form operational taxonomic units (OTUs) and then maps reads against the OTUs to
form a OTU by sample table. It depends on the folder structure created by the function Folders().
}
