% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseTaxonomy.R
\name{ParseTaxonomy}
\alias{ParseTaxonomy}
\title{ParseTaxonomy}
\usage{
ParseTaxonomy(pctThreshold = 97, lwrpctThreshold = 85,
  lwrcovpct = 65, blastoutput = NA, lineages = NA,
  returntable = TRUE)
}
\arguments{
\item{pctThreshold}{Percent identity match threashold for high quality assignments, 97-100 depending on marker}

\item{lwrpctThreshold}{Percent identity match threshold under which a hit is not considered}

\item{lwrcovpct}{Percent coverage of hit under which a hit is not considered}

\item{blastoutput}{The full directory to a text file with output from the BLAST standalone executable}

\item{lineages}{The full directory to a text file containing lineage information for each NCBI taxonomy number}

\item{returntable}{Return a table showing the assignments by catagory?}
}
\value{
Returns a data frame object containing assignments
}
\description{
This function parses taxonomy from a BLAST query to provide assignments at a given identity.
An example blast command to use to generate a dataset to use for assignbments might be blastn -query ../path/to/your/seqs.fasta -db nt -out ../path/to/your/resultsfolder/results.txt -outfmt '6 qseqid qlen slen qcovs qcovhsp sseqid bitscore score evalue pident qstart qend sstart send staxids sscinames' -num_alignments 25 -num_threads 1
}
