% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolNFilterReads.R
\name{PoolNFilterReads}
\alias{PoolNFilterReads}
\title{PoolNFilterReads}
\usage{
PoolNFilterReads(FastqMaxEE = 1, folderwfiles = "3.strippedreads",
  folderoutput = "4.pooledsamples", UsePrimerFile = FALSE,
  vsearchdest = "vsearch")
}
\arguments{
\item{FastqMaxEE}{Maximum expected error of pooled sequences}

\item{folderwfiles}{Folder containing fastq files to be processed. Default is 3.strippedreads.}

\item{folderoutput}{Output folder for processed reads. Default is 4.pooledsamples.}

\item{UsePrimerFile}{TRUE or FALSE, is a csv file provided containing primer information?}

\item{vsearchdest}{Specify the location of the vsearch executable if PATH not configured appropriately.}
}
\value{
None
}
\description{
This function pools and performs quality filtering on merged metabarcoding reads.
}
