% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrimerStrip.R
\name{PrimerStrip}
\alias{PrimerStrip}
\title{PrimerStrip}
\usage{
PrimerStrip(
  PrimerF = NA,
  PrimerR = NA,
  MinLen = NA,
  MaxLen = NA,
  UsePrimerFile = FALSE,
  folderwfiles = "2.mergedreads",
  folderoutput = "3.strippedreads",
  cutadaptdest = "cutadapt",
  ncores = 1
)
}
\arguments{
\item{PrimerF}{The 5-3 sequence of the forward primer. Ambiguous sequences are accepted (eg.N).}

\item{PrimerR}{The 5-3 sequence of the reverse primer. Ambiguous sequences are accepted (eg.N).}

\item{MinLen}{Minimum sequence length to be retained.}

\item{MaxLen}{Maximum sequence length to be retained.}

\item{UsePrimerFile}{TRUE or FALSE, is a csv file provided containing primer information?}

\item{folderwfiles}{A directory indicating files to be processed if not default folder.}

\item{folderoutput}{Output directory.}

\item{cutadaptdest}{Full directory for the cutadapt executable if not in PATH.}

\item{ncores}{Number of processsor cores to use.}
}
\value{
None
}
\description{
This function uses cutadapt to strip off non-biological nucleotides (adapters or primers) from fastq files. It depends on Cutadapt.
}
