% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenameFiles.R
\name{RenameFiles}
\alias{RenameFiles}
\title{RenameFiles}
\usage{
RenameFiles(SeqIDs, DesiredIDs, folderwfiles = "1.rawreads",
  usemetadatafile = FALSE)
}
\value{
None
}
\description{
This function renames .fq files according to a user supplied index.
}
\examples{
Folders()

}
