% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unzip.R
\name{Unzip}
\alias{Unzip}
\title{UnZip Files}
\usage{
Unzip(folderwfiles = "1.rawreads", unpigzlocation = "")
}
\arguments{
\item{folderwfiles}{Specify a folder other than '1.rawreads' to look for Gzip files.}

\item{unpigzlocation}{If the 'unpigz' executable in not in your path specify the location here.}
}
\value{
None
}
\description{
This function unzips gzip files. Illumina currently zips all raw sequencing
output in gzip. The fucntion looks for gzip files in the 1.rawreads folder,
alternatively another folder can be sepcified.
It relies on pigz so rock on over to 'https://zlib.net/pigz/' and
install a copy.
}
\examples{
\dontrun{Unzip()}

}
