% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenameFiles.R
\name{RenameFiles}
\alias{RenameFiles}
\title{RenameFiles}
\usage{
RenameFiles(
  SeqIDs,
  DesiredIDs,
  folderwfiles = "1.rawreads",
  usemetadatafile = FALSE
)
}
\arguments{
\item{SeqIDs}{A vector containing the filenames of the files to be renamed.}

\item{DesiredIDs}{A vector containing the desired names of the files to be renamed, in the same order as SeqIDs}

\item{folderwfiles}{A directory containign the files to be renamed, if not default directory.}

\item{usemetadatafile}{TRUE or FALSE, should the names in the metadata csv be used?}
}
\value{
None
}
\description{
This function renames .fq files according to a user supplied index.
}
